% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_dictionary}
\alias{data_dictionary}
\title{Data dictionary}
\format{
\subsection{\code{data_dictionary}}{

\describe{
\item{dataset}{Name of the dataset}
\item{column}{Name of the column}
\item{typeof}{Data type of the column}
\item{definition}{Description of what the column stands for}
\item{value}{Which values are allowed for the column}
...
}
}
}
\source{
internal
}
\usage{
data_dictionary
}
\description{
An overview of the output data sets generated by the package, their data types,
and the definitions of the variables.
}
\details{
For more details see the help vignette:
\code{vignette("data_dictionary", package = "workflow.multi.loanbook")}
}
\keyword{datasets}
