% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_loanbooks.R
\name{plot_aggregate_loanbooks}
\alias{plot_aggregate_loanbooks}
\title{plot_aggregate_loanbooks}
\usage{
plot_aggregate_loanbooks(config)
}
\arguments{
\item{config}{either a single string defining the path to a config YML file
or a list object that contains the appropriate config params}
}
\value{
`NULL` (called for side effects)
}
\description{
plot_aggregate_loanbooks
}
