% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\title{Make a sankey plot}
\usage{
plot_sankey(
  data,
  group_var,
  capitalise_node_labels = TRUE,
  save_png_to = NULL,
  png_name = "sankey.png",
  nodes_order_from_data = FALSE
)
}
\arguments{
\item{data}{data.frame. Should have the same format as output of
\code{prep_sankey()} and contain columns: \code{"middle_node"}, optionally
\code{"middle_node2"}, \code{"is_aligned"}, \code{"loan_size_outstanding"}, and any column
implied by \code{group_var}.}

\item{group_var}{Character. Vector of length 1. Variable to group by.}

\item{capitalise_node_labels}{Logical. Flag indicating if node labels should
be converted into better looking capitalised form.}

\item{save_png_to}{Character. Path where the output in png format should be
saved}

\item{png_name}{Character. File name of the output.}

\item{nodes_order_from_data}{Logical. Flag indicating if nodes order should
be determined by an algorithm (in case of big datasets often results in a
better looking plot) or should they be ordered based on data.}
}
\description{
Make a sankey plot
}
\examples{
# TODO
}
