% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_alignment_exposure.R,
%   R/prep_scatter_alignment_exposure.R
\name{plot_scatter_alignment_exposure}
\alias{plot_scatter_alignment_exposure}
\alias{prep_scatter_alignment_exposure}
\title{Plot alignment scatterplot}
\usage{
plot_scatter_alignment_exposure(
  data,
  floor_outliers,
  cap_outliers,
  group_var,
  currency
)

prep_scatter_alignment_exposure(
  data,
  year,
  region,
  scenario,
  group_var,
  exclude_groups = "benchmark"
)
}
\arguments{
\item{data}{data.frame. Holds net aggregated alignment metrics on the loan
book level. Must contain columns: \code{"scenario"}, \code{"region"},
\code{"sector"}, \code{"year"}, \code{"exposure_weighted_net_alignment"},
\code{"sum_loan_size_outstanding"} and any column implied by \code{group_var}.}

\item{floor_outliers}{Numeric. Floor which should be applied to the alignment
values in the data. Values smaller than floor are plotted on the border of
the plot.}

\item{cap_outliers}{Numeric. Cap which should be applied to the alignment
values in the data. Values bigger than cap are plotted on the border of the
plot.}

\item{group_var}{Character. Vector of length 1. A column to group by. Must be
available variables in \code{data}.}

\item{currency}{Character. Currency to display in the plot labels.}

\item{year}{Integer. Year on which \code{data} should be filtered.}

\item{region}{Character. Region to filter \code{data} data frame on.}

\item{scenario}{Character. Scenario to filter \code{data} data frame on.}

\item{exclude_groups}{Character. Character specifying any values from
\code{group_var} that should not be included in the analysis. This is useful to
remove benchmarks that are not meant to be compared at the same level.
Defaults to \code{"benchmark"}.}
}
\value{
object of type "ggplot"

data.frame
}
\description{
Plot alignment scatterplot
}
\examples{
# TODO
}
