% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter_animated.R,
%   R/prep_scatter_animated.R
\name{plot_scatter_animated}
\alias{plot_scatter_animated}
\alias{prep_scatter_animated}
\title{Plot alignment scatterplot}
\usage{
plot_scatter_animated(
  data,
  data_level = c("company", "group_var"),
  sector = NULL,
  scenario_source = NULL,
  scenario = NULL,
  region = NULL,
  title = NULL,
  subtitle = NULL,
  alignment_limit = NULL,
  cap_outliers = NULL,
  floor_outliers = NULL
)

prep_scatter_animated(
  data_bopo,
  data_net,
  data_level = c("group_var", "company"),
  sector,
  region,
  group_var,
  groups_to_plot = NULL
)
}
\arguments{
\item{data}{data.frame. Should have the same format as output of
\code{prep_scatter_animated()} and contain columns: 'name', 'buildout',
'phaseout', 'net' and 'year'.}

\item{data_level}{Character. Level of the plotted data. Can be \code{'group_var'} or
\code{'company'}.}

\item{sector}{Character. Sector to filter data on.}

\item{scenario_source}{Character. Scenario source to be used in the plot
caption.}

\item{scenario}{Character. Scenario name to be used in the plot caption.}

\item{region}{Character. Region to filter data on.}

\item{title}{Character. Custom title if different than default.}

\item{subtitle}{Character. Custom subtitle if different than default.}

\item{alignment_limit}{Numeric. Limit to be applied to the x- and y-axis
scales and to alignment values for colouring. By default the maximum
absolute alignment value from data is used.}

\item{cap_outliers}{Numeric. Cap which should be applied to the alignment
values in the data. Values bigger than cap are plotted on the border of the
plot.}

\item{floor_outliers}{Numeric. Floor which should be applied to the alignment
values in the data. Values smaller than floor are plotted on the border of
the plot.}

\item{data_bopo}{data.frame. Data containing buildout and phaseout alignment
values. Must contain columns: \code{'year'}, \code{'sector'}, \code{'region'},
\code{'direction'} and either \code{'name_abcd'} and \code{'alignment_metric'} or
\code{'exposure_weighted_net_alignment'} plus any column implied by \code{group_var}.}

\item{data_net}{data.frame. Data containing net alignment values. Must
contain columns: \code{group_var}, \code{'year'}, \code{'sector'}, \code{'region'}, \code{'direction'} and
either \code{'name_abcd'} and \code{'alignment_metric'} or \code{'exposure_weighted_net_alignment'}.}

\item{group_var}{Character. Vector of length 1. Variable to group by.}

\item{groups_to_plot}{Character vector. Groups to filter on.}
}
\value{
object of type "plotly"

data.frame
}
\description{
Plot alignment scatterplot
}
\examples{
# TODO
}
