% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeline.R
\name{plot_timeline}
\alias{plot_timeline}
\title{Plot alignment timeline}
\usage{
plot_timeline(
  data,
  sector = NULL,
  scenario_source = NULL,
  scenario = NULL,
  region = NULL,
  group_var = NULL,
  title = NULL,
  subtitle = NULL,
  alignment_limits = NULL
)
}
\arguments{
\item{data}{data.frame Should have the same format as output of
\code{prep_timeline()} and contain columns: \code{'direction'}, \code{'year'},
\code{'exposure_weighted_net_alignment'}, and any column implied by \code{group_var}.}

\item{sector}{Character. Sector name to be used in the plot title.}

\item{scenario_source}{Character. Scenario source to be used in the plot
caption.}

\item{scenario}{Character. Scenario name to be used in the plot caption.}

\item{region}{Character. Region to be used in the plot caption.}

\item{group_var}{Character. Vector of length 1. Variable to group by.}

\item{title}{Character. Custom title if different than default.}

\item{subtitle}{Character. Custom subtitle if different than default.}

\item{alignment_limits}{Numeric vector of size 2. Limits to be applied to
alignment values for colouring. By default maximum absolute value of
'exposure_weighted_net_alignment' is used.}
}
\value{
object of type "ggplot"
}
\description{
Plot alignment timeline
}
\examples{
# TODO
}
