% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_sankey.R
\name{prep_sankey}
\alias{prep_sankey}
\title{Prepare data to plot using \code{plot_sankey()}}
\usage{
prep_sankey(
  data_alignment,
  region,
  year,
  group_var,
  middle_node,
  middle_node2 = NULL
)
}
\arguments{
\item{data_alignment}{data.frame. Holds aggregated alignment metrics per
company for tms sectors. Must contain columns: \code{"name_abcd"},
\code{"sector"} and any column implied by \code{group_var}.}

\item{region}{Character. Region to filter \code{data_alignment} data frame on.}

\item{year}{Integer. Year on which \code{data_alignment} should be filtered.}

\item{group_var}{Character. Vector of length 1. Variable to group by.}

\item{middle_node}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}

\item{middle_node2}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}
}
\value{
data.frame
}
\description{
Prepare data to plot using \code{plot_sankey()}
}
\examples{
# TODO
}
