% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_scatter_alignment_exposure.R
\name{prep_scatter_alignment_exposure}
\alias{prep_scatter_alignment_exposure}
\title{Prepare data to plot using \code{plot_scatter_alignment_exposure()}}
\usage{
prep_scatter_alignment_exposure(
  data,
  year,
  region,
  scenario,
  group_var,
  exclude_groups = "benchmark"
)
}
\arguments{
\item{data}{data.frame. Holds net aggregated alignment metrics on the loan
book level. Must contain columns: \code{"scenario"}, \code{"region"},
\code{"sector"}, \code{"year"}, \code{"exposure_weighted_net_alignment"},
\code{"sum_loan_size_outstanding"} and any column implied by \code{group_var}.}

\item{year}{Integer. Year on which \code{data} should be filtered.}

\item{region}{Character. Region to filter \code{data} data frame on.}

\item{scenario}{Character. Scenario to filter \code{data} data frame on.}

\item{group_var}{Character. Vector of length 1. A column to group by. Must be
available variables in \code{data}.}

\item{exclude_groups}{Character. Character specifying any values from
\code{group_var} that should not be included in the analysis. This is useful to
remove benchmarks that are not meant to be compared at the same level.
Defaults to \code{"benchmark"}.}
}
\value{
data.frame
}
\description{
Prepare data to plot using \code{plot_scatter_alignment_exposure()}
}
\examples{
# TODO
}
