% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_pacta.R
\name{run_pacta}
\alias{run_pacta}
\title{run_pacta}
\usage{
run_pacta(config)
}
\arguments{
\item{config}{either a single string defining the path to a config YML file
or a list object that contains the appropriate config params}
}
\value{
\code{NULL} (called for side effects)
}
\description{
run_pacta
}
