% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_aggregate_loanbooks.R
\name{plot_aggregate_loanbooks}
\alias{plot_aggregate_loanbooks}
\title{plot_aggregate_loanbooks}
\usage{
plot_aggregate_loanbooks()
}
\value{
\code{NULL} (called for side effects)
}
\description{
plot_aggregate_loanbooks
}
