% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sankey.R, R/prep_sankey.R
\name{plot_sankey}
\alias{plot_sankey}
\alias{prep_sankey}
\title{Make a sankey plot}
\usage{
plot_sankey(
  data,
  group_var,
  capitalise_node_labels = TRUE,
  save_png_to = NULL,
  png_name = "sankey.png",
  nodes_order_from_data = FALSE
)

prep_sankey(
  data_alignment,
  region,
  year,
  group_var,
  middle_node,
  middle_node2 = NULL
)
}
\arguments{
\item{data}{data.frame. Should have the same format as output of
\code{prep_sankey()} and contain columns: \code{"middle_node"}, optionally
\code{"middle_node2"}, \code{"is_aligned"}, \code{"loan_size_outstanding"}, and any column
implied by \code{group_var}.}

\item{group_var}{Character. Vector of length 1. Variable to group by.}

\item{capitalise_node_labels}{Logical. Flag indicating if node labels should
be converted into better looking capitalised form.}

\item{save_png_to}{Character. Path where the output in png format should be
saved}

\item{png_name}{Character. File name of the output.}

\item{nodes_order_from_data}{Logical. Flag indicating if nodes order should
be determined by an algorithm (in case of big datasets often results in a
better looking plot) or should they be ordered based on data.}

\item{data_alignment}{data.frame. Holds aggregated alignment metrics per
company for tms sectors. Must contain columns: \code{"name_abcd"},
\code{"sector"} and any column implied by \code{group_var}.}

\item{region}{Character. Region to filter \code{data_alignment} data frame on.}

\item{year}{Integer. Year on which \code{data_alignment} should be filtered.}

\item{middle_node}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}

\item{middle_node2}{Character. Column specifying the middle nodes to be
plotted in sankey plot. Must be present in \code{data_alignment}.}
}
\value{
data.frame
}
\description{
Make a sankey plot
}
\examples{
# TODO
}
